+++ tests/test_app/TestApp/Plugin/MeCmsExample/config/bootstrap.php
@@ -1,0 +1,6 @@
+<?php
+
+use Cake\Core\Configure;
+use MeCms\Example\View\Helper\ExampleMenuHelper;
+
+Configure::write('MeCmsExample.MenuHelpers', ExampleMenuHelper::class);