+++ tests/test_app/TestApp/Plugin/MeCmsExample/src/View/Helper/ExampleMenuHelper.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace MeCms\Example\View\Helper;
+
+use MeCms\View\Helper\AbstractMenuHelper;
+
+class ExampleMenuHelper extends AbstractMenuHelper
+{
+    public function getLinks(): array
+    {
+        return [];
+    }
+
+    public function getOptions(): array
+    {
+        return [];
+    }
+
+    public function getTitle(): string
+    {
+        return '';
+    }
+}